UI: Fix problem in Component when previous type is already loaded and route is changed
Component#addTypeItemStateVariable allowed an async function (props) => id as parameter id.
In Component#_loadAsyncVarItem the result from that function was awaited before setting
state variable loadingAsync to true. The await resulted in a yield which did let the
render function to run before loadingAsync was set. This resulted in rendering one component
with data for another...