UI: Fix problem in TAView when item subscription not setup
If navigating from route without a path starting with :xx to a
path starting with :xx TAView item subscription is not setup
since the constructor isn't run again. Removed if-statement
in constructor that conditonally enables the "item" state variable
and instead add an enabled callback to Component#addTypeItemStateVariable
that is run when item is loaded instead.