Add active directory (AD) backend to userrepo (#306)
Active directory backend that uses npm module activedirectory2. If user is not found in userRepo it will try to find it in any of the defined userRepo backends configured. 

The AD backend requires a url, baseDN and it is possible to select an default user policy of the ones created so that new users will receive that policy automatically.